Simplified SimpleExtension. Moved manual registration example to ManualExtension.
