+++ Sources/ManualExtension/ManualExtension.swift
@@ -9,6 +9,7 @@ import SwiftGodot
 
 var sequence = 0
 
+
 @Godot
 class Rigid: RigidBody2D {
     override func _integrateForces(state: PhysicsDirectBodyState2D?) {
@@ -19,86 +20,15 @@ class Rigid: RigidBody2D {
     }
 }
 
-@Godot
-class SwiftSprite: Sprite2D {
-    var time_passed: Double = 0
-    var count: Int = 0
-    
-    #signal("picked_up_item", arguments: ["kind": String.self, "isGroovy": Bool.self, "count": Int.self])
-    #signal("scored")
-    #signal("lives_changed", arguments: ["count": Int.self])
-    
-    @Callable
-    public func computeGodot (x: String, y: Int) -> Double {
-        return 1.0
-    }
-
-    @Callable
-    public func wink () {
-        print ("Wink")
-    }
-    
-    @Callable
-    public func computerSimple (_ x: Int, _ y: Int) -> Double {
-        return Double (x + y)
-    }
-    
-    @Callable 
-    func returnNullable () -> String? {
-        let x: Variant = Variant (1)
-        if let y: Resource = x.asObject () {
-            print ("Y is = \(y)")
-        }
-        return nil
-    }
-    
-    @Export var resource: Resource?
-    @Export(.dir) var directory: String?
-    @Export(.file, "*txt") var file: String?
-    @Export var demo: String = "demo"
-    @Export var food: String = "none"
-    
-    var x: Rigid?
-    
-    override func _process (delta: Double) {
-        time_passed += delta
-    
-        if x == nil {
-            self.x = Rigid()
-        }
-        guard let imageVariant = ProjectSettings.getSetting(name: "shader_globals/heightmap", defaultValue: Variant(-1)) else {
-            return
-        }
-        
-        GD.print("Found this value IMAGE: \(imageVariant.gtype) variant: \(imageVariant) desc: \(imageVariant.description)")
-        
-        let dict2: GDictionary? = GDictionary(imageVariant)
-       GD.print("dictionary2: \(String(describing: dict2)) \(dict2?["type"]?.description ?? "no type") \(dict2?["value"]?.description ?? "no value")")
-        
-        // part b
-        if let result = dict2?.get(key: Variant("type"), default: Variant(-1)) {
-            let value = String(result) ?? "No Result"
-            GD.print("2 Found this value \(value)")
-        }
-        
-        let lerp = Double(0.1).lerp(to: 10, weight: 1)
-        print ("Lerp result from 0.1 to 10 weight:1 => \(lerp)")
-        let newPos = Vector2(x: Float (10 + (10 * sin(time_passed * 2.0))),
-                             y: Float (10.0 + (10.0 * cos(time_passed * 1.5))))
-        
-        self.position = newPos
-    }
-}
-
 // This shows how to register methods and properties manually
-class SwiftSprite2: Sprite2D {
+class SwiftSprite: Sprite2D {
     var time_passed: Double
     var count: Int
     
     // This is a class initializer, must be invoked from all of your constructors to register the various
     // features of this class with the Godot engine.
     static var initClass: Void = {
-        let classInfo = ClassInfo<SwiftSprite2> (name: "SwiftSprite2")
+        let classInfo = ClassInfo<SwiftSprite> (name: "SwiftSprite")
         
         classInfo.addPropertyGroup(name: "Miguel's Demo", prefix: "demo_")
         
@@ -110,8 +40,8 @@ class SwiftSprite2: Sprite2D {
                      hintStr: "Some kind of food",
                      usage: .default)
         ]
-        classInfo.registerMethod(name: "demo_set_favorite_food", flags: .default, returnValue: nil, arguments: foodArgs, function: SwiftSprite2.demoSetFavoriteFood)
-        classInfo.registerMethod(name: "demo_get_favorite_food", flags: .default, returnValue: foodArgs [0], arguments: [], function: SwiftSprite2.demoGetFavoriteFood)
+        classInfo.registerMethod(name: "demo_set_favorite_food", flags: .default, returnValue: nil, arguments: foodArgs, function: SwiftSprite.demoSetFavoriteFood)
+        classInfo.registerMethod(name: "demo_get_favorite_food", flags: .default, returnValue: foodArgs [0], arguments: [], function: SwiftSprite.demoGetFavoriteFood)
         
         let foodProp = PropInfo (propertyType: .string,
                                  propertyName: "demo_favorite_food",
@@ -123,14 +53,14 @@ class SwiftSprite2: Sprite2D {
     }()
     
     required init (nativeHandle: UnsafeRawPointer) {
-        _ = SwiftSprite2.initClass
+        _ = SwiftSprite.initClass
         time_passed = 0
         count = sequence
         super.init (nativeHandle: nativeHandle)
     }
     
     required init () {
-        _ = SwiftSprite2.initClass
+        _ = SwiftSprite.initClass
         count = sequence
         sequence += 1
         time_passed = 0
@@ -191,22 +121,22 @@ class SwiftSprite2: Sprite2D {
     }
 }
 
+/// Setup 
 func setupScene (level: GDExtension.InitializationLevel) {
     if level == .scene {
-        register(type: SwiftSprite.self)
-        register(type: SwiftSprite2.self)
         register(type: Rigid.self)
+        register(type: SwiftSprite.self)
     }
 }
 
-// Set the swift.gdextension's entry_symbol to "swift_entry_point
+/// Manually defined entry point for the extension.
 @_cdecl("swift_entry_point")
 public func swift_entry_point(
     godotGetProcAddr: OpaquePointer?,
     libraryPtr: OpaquePointer?,
     extensionPtr: OpaquePointer?) -> UInt8
 {
-    print ("SwiftSprite: Starting up")
+    print ("ManualExtension: Starting up")
     guard let godotGetProcAddr, let libraryPtr, let extensionPtr else {
         return 0
     }
