+++ Sources/ManualExtension/README.md
@@ -1,0 +1,14 @@
+#  ManualExtension
+
+This example extension shows how you can write custom registration code
+for your extension, to do more complicated things.
+
+For a simpler example, see SimpleExtension.
+
+The code is contained in ManualExtension.swift, the package definition
+is contained by ../Package.swift and you must copy the 
+resulting binary (ManualExtension) into the bin directory 
+of your Godot project and place the ManualExtension.gdextension file
+in the Godot project root.
+
+
