+++ Sources/SimpleExtension/README.md
@@ -3,10 +3,10 @@
 This simple extension shows how you can create a class that
 is exposed to Godot and how you bootstrap it.
 
-The code is contained in Demo.swift, the package definition
+The code is contained in SimpleExtension.swift, the package definition
 is contained by ../Package.swift and you must copy the 
-resulting binary (SimpleExtnesion) into the bin directory 
-of your Godot project and place the SwiftSprite.gdextension file
+resulting binary (SimpleExtension) into the bin directory 
+of your Godot project and place the SimpleExtension.gdextension file
 in the Godot project root.
